+++ src/main/java/com/example/ajaxonboardingproject/service/mapper/MovieSessionMapper.kt
@@ -23,10 +23,10 @@ class MovieSessionMapper(
 
     override fun mapToDto(model: MovieSession): MovieSessionResponseDto {
         return MovieSessionResponseDto(
-            movieSessionId = model.id,
-            cinemaHallId = model.cinemaHall.id,
+            movieSessionId = model.id.toLong(),
+            cinemaHallId = model.cinemaHall.id.toLong(),
             movieTitle = model.movie.title,
-            movieId = model.movie.id,
+            movieId = model.movie.id.toLong(),
             showTime = model.showTime
         )
     }
